Check app.login_manager for Flask-Login
Instead of checking for flask_login in sys.modules, check
that current_app.login_manager has been populated. It's
nicer, and also a better check (because it will only be true
if Flask-Login is enabled for the current app, not enough
if it's just installed in the current Python env).